Instrument code with tracing of parameter/block growth, and unitarity progress
The first change involves the atlas program, the second scripts hermitian.at and
basic.at; the latter to reverse processing of terms in |map| of a function of
type |(Param->ParamPol)| over a |ParamPol|, because it turns out the lowest
height term processed in |c_form_irreducuble| has a very long |full_deform|.

Since we now check for user interrupts in the repr.cpp code, we can no longer
link the Fokko program successfully. That program being the least of our
concerns for this patch, this linking has simply been disabled in the Makefile.